+++ homekit_bridge_manager/run.sh
@@ -4,6 +4,7 @@ set -euo pipefail
 
 export REFRESH_SECONDS="$(bashio::config 'refresh_seconds')"
 export LOG_LEVEL="$(bashio::config 'log_level')"
+export ASSUME_NOT_IN_HOMEKIT="$(bashio::config 'assume_not_in_homekit | join(",")')"
 export HA_CONFIG_DIR="/homeassistant"
 export PORT="8099"